refactor: plan current sweeps in process through infx
通过 infx.matrix.plan 统一规划当前扫描，避免重复启动生成器并加载配置；保留历史生成隔离和 CLI 行为，使用真实配置对照测试验证兼容性。
