+++ infx/matrix/__init__.py
@@ -1,1 +1,5 @@
 """Benchmark configuration schemas and matrix generation."""
+
+MASTER_CONFIGS = ["configs/amd-master.yaml", "configs/nvidia-master.yaml"]
+RUNNER_CONFIG = "configs/runners.yaml"
+GENERATE_SWEEPS_PY_SCRIPT = "utils/matrix_logic/generate_sweep_configs.py"
