+++ utils/test_process_changelog.py
@@ -4,7 +4,6 @@ import json
 import shutil
 import subprocess
 import sys
-from contextlib import nullcontext
 from pathlib import Path
 from types import SimpleNamespace
 
@@ -136,13 +135,6 @@ def _fixed_matrix_row(
         "run-eval": False,
         "eval-only": False,
     } | ({"duration": duration} if duration is not None else {})
-
-
-def _scenario_values(command):
-    if "--scenario-type" not in command:
-        return []
-    index = command.index("--scenario-type") + 1
-    return command[index:]
 
 
 def test_trim_conc_supports_nested_backend_metadata():
@@ -528,528 +520,59 @@ def test_append_only_scope_allows_range_to_list_expansion():
     )
 
 
-@pytest.mark.parametrize("trim", [False, True])
-def test_append_only_main_runs_only_added_points_and_skips_evals(
-    monkeypatch,
-    capsys,
-    trim,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Add one concurrency point without rerunning the curve
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  append-only: true
-"""
-    base_rows = [_fixed_matrix_row(4)]
-    head_rows = [*base_rows, _fixed_matrix_row(8)]
-    commands = []
-
-    monkeypatch.setattr(process_changelog, "get_added_lines", lambda *_: added_yaml)
-    monkeypatch.setattr(
-        process_changelog,
-        "generation_inputs_at_ref",
-        lambda *_: nullcontext(
-            process_changelog.GenerationInputs(
-                config_files=["base-nvidia.yaml", "base-amd.yaml"],
-                generator_script="base-generate-sweep-configs.py",
-                runner_config="base-runners.yaml",
-            )
-        ),
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {"image": "vllm/vllm-openai:v0.16.0"}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        rows = base_rows if "base-nvidia.yaml" in command else head_rows
-        return SimpleNamespace(stdout=json.dumps(rows))
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml", *(["--trim-conc"] if trim else []),
-    ])
-
-    process_changelog.main()
-
-    output = json.loads(capsys.readouterr().out)
-    assert [row["conc"] for row in output["single_node"]["8k1k"]] == [8]
-    assert len(output["single_node"]["8k1k"][0]["recipe-fingerprint"]) == 64
-    assert output["evals"] == []
-    assert output["changelog_metadata"]["entries"][0]["append-only"] is True
-    assert len(commands) == 2
-    assert commands[0][1] == process_changelog.GENERATE_SWEEPS_PY_SCRIPT
-    assert commands[1][1] == "base-generate-sweep-configs.py"
-    assert commands[0][commands[0].index("--runner-config") + 1] == "configs/runners.yaml"
-    assert commands[1][commands[1].index("--runner-config") + 1] == "base-runners.yaml"
-
-
-def test_cli_evals_only_generates_agentic_eval(
-    monkeypatch,
-    capsys,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Agentic-only work with the evals-only PR modifier
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - agentic-coding
-"""
-    commands = []
-
-    monkeypatch.setattr(
-        process_changelog,
-        "get_added_lines",
-        lambda *_: added_yaml,
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        return SimpleNamespace(stdout="[]")
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-        "--evals-only",
-    ])
-
-    process_changelog.main()
-
-    assert len(commands) == 1
-    assert "--evals-only" in commands[0]
-    assert "--no-evals" not in commands[0]
-    assert _scenario_values(commands[0]) == ["agentic-coding"]
-    json.loads(capsys.readouterr().out)
-
-
-def test_cli_all_evals_generates_agentic_eval(
-    monkeypatch,
-    capsys,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Agentic-only work with the all-evals PR modifier
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - agentic-coding
-"""
-    commands = []
-
-    monkeypatch.setattr(
-        process_changelog,
-        "get_added_lines",
-        lambda *_: added_yaml,
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        return SimpleNamespace(stdout="[]")
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-        "--all-evals",
-    ])
-
-    process_changelog.main()
-
-    assert len(commands) == 2
-    assert "--no-evals" in commands[0]
-    assert _scenario_values(commands[0]) == ["agentic-coding"]
-    assert "--evals-only" in commands[1]
-    assert "--all-evals" in commands[1]
-    assert _scenario_values(commands[1]) == ["agentic-coding"]
-    json.loads(capsys.readouterr().out)
-
-
-def test_all_evals_takes_precedence_for_duplicate_configs(
-    monkeypatch,
-    capsys,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Regular benchmark entry appears first
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-
-- config-keys:
-    - test-config
-  description:
-    - Expand the same config to all evals
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  all-evals: true
-"""
-    commands = []
-
-    monkeypatch.setattr(
-        process_changelog,
-        "get_added_lines",
-        lambda *_: added_yaml,
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        return SimpleNamespace(stdout="[]")
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-    ])
-
-    process_changelog.main()
-
-    assert len(commands) == 2
-    assert "--all-evals" in commands[0]
-    assert "--evals-only" in commands[0]
-    assert "--no-evals" in commands[1]
-    json.loads(capsys.readouterr().out)
-
-
-def test_disjoint_scenario_entries_for_same_config_are_not_deduplicated(
-    monkeypatch,
-    capsys,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Fixed sequence jobs
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - fixed-seq-len
-
-- config-keys:
-    - test-config
-  description:
-    - Agentic jobs
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - agentic-coding
-"""
-    commands = []
-
-    monkeypatch.setattr(
-        process_changelog,
-        "get_added_lines",
-        lambda *_: added_yaml,
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        return SimpleNamespace(stdout="[]")
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-    ])
-
-    process_changelog.main()
-
-    assert len(commands) == 4
-    assert "--no-evals" in commands[0]
-    assert _scenario_values(commands[0]) == ["fixed-seq-len"]
-    assert "--evals-only" in commands[1]
-    assert _scenario_values(commands[1]) == ["fixed-seq-len"]
-    assert "--no-evals" in commands[2]
-    assert _scenario_values(commands[2]) == ["agentic-coding"]
-    assert "--evals-only" in commands[3]
-    assert _scenario_values(commands[3]) == ["agentic-coding"]
-    json.loads(capsys.readouterr().out)
-
-
-def test_agentic_only_all_evals_does_not_suppress_later_fixed_evals(
-    monkeypatch,
-    capsys,
-):
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Agentic-only all-evals entry
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - agentic-coding
-  all-evals: true
-
-- config-keys:
-    - test-config
-  description:
-    - Fixed sequence jobs
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-  scenario-type:
-    - fixed-seq-len
-"""
-    commands = []
-
-    monkeypatch.setattr(
-        process_changelog,
-        "get_added_lines",
-        lambda *_: added_yaml,
-    )
-    monkeypatch.setattr(
-        process_changelog,
-        "load_config_files",
-        lambda _: {"test-config": {}},
-    )
-
-    def fake_run(command, **kwargs):
-        commands.append(command)
-        return SimpleNamespace(stdout="[]")
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py",
-        "--base-ref", "base",
-        "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-    ])
-
-    process_changelog.main()
-
-    assert len(commands) == 3
-    assert "--evals-only" in commands[0]
-    assert "--all-evals" in commands[0]
-    assert _scenario_values(commands[0]) == ["agentic-coding"]
-    assert "--no-evals" in commands[1]
-    assert _scenario_values(commands[1]) == ["fixed-seq-len"]
-    assert "--evals-only" in commands[2]
-    assert "--all-evals" not in commands[2]
-    assert _scenario_values(commands[2]) == ["fixed-seq-len"]
-    json.loads(capsys.readouterr().out)
-
-
-def test_eval_rows_split_into_fixed_and_agentic_buckets(
-    monkeypatch,
-    capsys,
-):
-    """Realistic eval rows must pass final validation and land in the bucket
-    matching their dispatch job: fixed-seq-len rows in `evals`, agentic
-    GSM8K rows in `agentic_evals`."""
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Mixed fixed-seq-len and agentic eval selection
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-"""
-    common = {
-        "image": "vllm/vllm-openai:v0.11.0",
-        "model": "deepseek-ai/DeepSeek-V4-Pro", "model-prefix": "dsv4",
-        "precision": "fp4", "framework": "vllm", "spec-decoding": "mtp",
-        "runner": "cluster:b300-nv", "tp": 8, "pp": 1, "dcp-size": 1,
-        "pcp-size": 1, "ep": 8, "dp-attn": True, "conc": 224,
-        "run-eval": True, "eval-only": True,
-    }
-    fixed_eval_row = {
-        **common, "isl": 8192, "osl": 1024, "max-model-len": 10240,
-        "disagg": False, "exp-name": "fixed_eval",
-    }
-    agentic_eval_row = {
-        **common, "kv-offloading": "none", "total-cpu-dram-gb": 0,
-        "duration": 3600, "scenario-type": "agentic-coding",
-        "exp-name": "agentic_eval",
-    }
-
-    monkeypatch.setattr(
-        process_changelog, "get_added_lines", lambda *_: added_yaml)
-    monkeypatch.setattr(
-        process_changelog, "load_config_files", lambda _: {"test-config": {}})
-
-    def fake_run(command, **kwargs):
-        is_evals = "--evals-only" in command
-        rows = [fixed_eval_row, agentic_eval_row] if is_evals else []
-        return SimpleNamespace(stdout=json.dumps(rows))
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py", "--base-ref", "base", "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-    ])
-
-    process_changelog.main()
-
-    output = json.loads(capsys.readouterr().out)
-    assert [r["exp-name"] for r in output["evals"]] == ["fixed_eval"]
-    assert [r["exp-name"] for r in output["agentic_evals"]] == ["agentic_eval"]
-    assert output["multinode_evals"] == []
-
-
-def test_eval_rows_split_into_multinode_fixed_and_agentic_buckets(
-    monkeypatch,
-    capsys,
-):
-    """Multi-node eval rows must split the same way single-node rows do:
-    fixed-seq-len rows in `multinode_evals`, agentic (SWE-bench) rows in
-    `multinode_agentic_evals`."""
-    added_yaml = """
-- config-keys:
-    - test-config
-  description:
-    - Mixed multi-node fixed-seq-len and agentic eval selection
-  pr-link: https://github.com/SemiAnalysisAI/InferenceX/pull/1
-"""
-    common = {
-        "image": "lmsysorg/sglang-rocm:v0.5.15", "model": "deepseek-ai/DeepSeek-V4-Pro",
-        "model-prefix": "dsv4", "precision": "fp4", "framework": "sglang-disagg",
-        "spec-decoding": "none", "runner": "cluster:mi355x-amds",
-        "node-count": 2,
-        "prefill": {"num-worker": 1, "tp": 8, "ep": 1, "dp-attn": False},
-        "decode": {"num-worker": 1, "tp": 8, "ep": 1, "dp-attn": False},
-        "disagg": True, "kv-p2p-transfer": "mori",
-        "run-eval": True, "eval-only": True,
-    }
-    multinode_fixed_eval_row = {
-        **common, "isl": 8192, "osl": 1024, "max-model-len": 10240,
-        "conc": [64], "eval-conc": 64, "exp-name": "multinode_fixed_eval",
-    }
-    multinode_agentic_eval_row = {
-        **common, "kv-offloading": "dram",
-        "kv-offload-backend": {"name": "hicache"},
-        "total-cpu-dram-gb": 2399, "duration": 3600,
-        "scenario-type": "agentic-coding",
-        "conc": [32], "eval-conc": 32, "exp-name": "multinode_agentic_eval",
-    }
-
-    monkeypatch.setattr(
-        process_changelog, "get_added_lines", lambda *_: added_yaml)
-    monkeypatch.setattr(
-        process_changelog, "load_config_files", lambda _: {"test-config": {}})
-
-    def fake_run(command, **kwargs):
-        is_evals = "--evals-only" in command
-        rows = (
-            [multinode_fixed_eval_row, multinode_agentic_eval_row]
-            if is_evals else []
-        )
-        return SimpleNamespace(stdout=json.dumps(rows))
-
-    monkeypatch.setattr(subprocess, "run", fake_run)
-    monkeypatch.setattr(sys, "argv", [
-        "process_changelog.py", "--base-ref", "base", "--head-ref", "head",
-        "--changelog-file", "perf-changelog.yaml",
-    ])
-
-    process_changelog.main()
-
-    output = json.loads(capsys.readouterr().out)
-    assert [r["exp-name"] for r in output["multinode_evals"]] == ["multinode_fixed_eval"]
-    assert [r["exp-name"] for r in output["multinode_agentic_evals"]] == ["multinode_agentic_eval"]
-    assert output["evals"] == []
-    assert output["agentic_evals"] == []
+@pytest.fixture
+def planning_repo(tmp_path, monkeypatch):
+    """Real CLI/config/generator wiring with small, independent input recipes."""
+    source = Path(__file__).resolve().parents[1]
+    for directory in ("utils/matrix_logic", "infx"):
+        shutil.copytree(source / directory, tmp_path / directory)
+    (tmp_path / "configs").mkdir()
+    (tmp_path / "configs/amd-master.yaml").write_text("{}\n")
+    runners = {"labels": {"cluster:fixture": ["node-a"]}, "hardware": {
+        "cluster:fixture": {"gpus-per-node": 8, "available-cpu-dram-mib": 1024000},
+    }}
+    (tmp_path / "configs/runners.yaml").write_text(yaml.safe_dump(runners))
+    master = {}
+    for key, multinode in (("single", False), ("multi", True)):
+        shape = ({role: {"num-worker": 1, "tp": 8, "ep": 1, "dp-attn": False}
+                  for role in ("prefill", "decode")} if multinode else {"tp": 8})
+        master[key] = {
+            "image": "example/image:stable", "model": key, "model-prefix": "dsr1",
+            "precision": "fp8", "framework": "sglang", "runner": "cluster:fixture",
+            "multinode": multinode, "disagg": multinode,
+            **({"kv-p2p-transfer": "nixl"} if multinode else {}),
+            "scenarios": {
+                "fixed-seq-len": [{"isl": 8192, "osl": 1024, "search-space": [
+                    {**shape, "conc-list": [16, 32, 64]},
+                ]}],
+                "agentic-coding": [{"search-space": [
+                    {**shape, "conc-list": [16, 32], **({} if multinode else {"kv-offloading": "none"})},
+                ]}],
+            },
+        }
+    (tmp_path / "configs/nvidia-master.yaml").write_text(yaml.safe_dump(master, sort_keys=False))
+    monkeypatch.chdir(tmp_path)
+    return tmp_path, master, runners
 
 
 @pytest.fixture
-def changelog_run(monkeypatch, capsys):
-    """Exercise main and final validation; replace only Git/config/subprocess I/O."""
-    def run(entries, cli_flags=(), generated=None, master=None):
-        entries = [{"config-keys": ["config-a"], "description": ["Controlled change"],
+def changelog_run(planning_repo, monkeypatch, capsys):
+    def run(entries, cli_flags=()):
+        entries = [{"config-keys": ["single"], "description": ["Controlled change"],
                     "pr-link": "https://github.com/SemiAnalysisAI/InferenceX/pull/1",
                     **entry} for entry in entries]
-        commands = []
         monkeypatch.setattr(process_changelog, "get_added_lines", lambda *_: json.dumps(entries))
-        monkeypatch.setattr(process_changelog, "load_config_files",
-                            lambda _: master if master is not None else {"config-b": {}, "config-a": {}})
         monkeypatch.setattr(sys, "argv", ["process_changelog.py", "--base-ref", "base",
                             "--head-ref", "head", "--changelog-file", "perf-changelog.yaml", *cli_flags])
-
-        def generate(command, **kwargs):
-            commands.append(command)
-            result = generated(command) if generated else []
-            if isinstance(result, subprocess.CalledProcessError):
-                if kwargs.get("check"):
-                    raise result
-                result = []
-            return SimpleNamespace(stdout=result if isinstance(result, str) else json.dumps(result))
-
-        monkeypatch.setattr(subprocess, "run", generate)
         process_changelog.main()
-        return json.loads(capsys.readouterr().out), commands
+        captured = capsys.readouterr()
+        assert captured.err == ""
+        return json.loads(captured.out)
     return run
 
 
-@pytest.mark.parametrize("skip", [False, True])
-def test_no_evals_preserves_throughput_and_metadata(changelog_run, skip):
-    output, commands = changelog_run(
-        [{"no-evals": skip}], generated=lambda _: [_fixed_matrix_row(8)],
-    )
-    assert [row["conc"] for row in output["single_node"]["8k1k"]] == [8]
-    assert len(output["evals"]) == (0 if skip else 1)
-    assert output["agentic_evals"] == []
-    assert output["multinode_evals"] == []
-    assert output["multinode_agentic_evals"] == []
-    assert len(commands) == (1 if skip else 2)
-    assert output["changelog_metadata"]["entries"][0]["no-evals"] is skip
-
-
-def test_no_evals_does_not_suppress_another_entry(changelog_run):
-    _, commands = changelog_run([
-        {"config-keys": ["config-a"], "no-evals": True},
-        {"config-keys": ["config-b"]},
-    ])
-    eval_commands = [command for command in commands if "--evals-only" in command]
-    assert len(eval_commands) == 1
-    assert eval_commands[0][eval_commands[0].index("--config-keys") + 1] == "config-b"
-
-
-@pytest.mark.parametrize("flags", [{"evals-only": True}, {"all-evals": True},
-                                   {"eval-min-prefill-ep": 2}])
-def test_no_evals_rejects_conflicting_entry_options(changelog_run, flags):
-    with pytest.raises(ValueError, match="no-evals cannot be combined"):
-        changelog_run([{"no-evals": True, **flags}])
-
-
-@pytest.mark.parametrize("flag", ["--all-evals", "--evals-only"])
-def test_no_evals_rejects_conflicting_pr_modifiers(changelog_run, flag):
-    with pytest.raises(ValueError, match="no-evals entries cannot use"):
-        changelog_run([{"no-evals": True}], [flag])
-
-
-# The table is the contract: CLI expansion preserves throughput, whereas an
-# entry requesting all evals is eval-only. Trimming affects throughput alone.
+# Hand-worked policy: entry all-evals suppresses throughput; the CLI modifier
+# expands evals while retaining throughput. Trimming affects throughput alone.
 @pytest.mark.parametrize("cli_flags,expected_modes", [
     ([], ("benchmark-subset", "subset", "all", "all")),
     (["--all-evals"], ("benchmark-all", "all", "all", "all")),
@@ -1062,97 +585,112 @@ def test_no_evals_rejects_conflicting_pr_modifiers(changelog_run, flag):
 ])
 @pytest.mark.parametrize("trim", [False, True])
 def test_cli_entry_mode_truth_table(changelog_run, cli_flags, expected_modes, entry_flags, mode_index, trim):
-    def generate(command):
-        concs = [8, 4] if "--no-evals" in command or "--all-evals" in command else [4]
-        return [_fixed_matrix_row(conc) for conc in concs]
-
-    output, commands = changelog_run([entry_flags], cli_flags + (["--trim-conc"] if trim else []), generate)
+    output = changelog_run([entry_flags], cli_flags + (["--trim-conc"] if trim else []))
     mode = expected_modes[mode_index]
-    expected_benchmarks = ([4] if trim else [8, 4]) if mode.startswith("benchmark-") else []
-    assert [row["conc"] for row in output["single_node"].get("8k1k", [])] == expected_benchmarks
-    assert [row["conc"] for row in output["evals"]] == ([8, 4] if mode.endswith("all") else [4])
-    assert ["--no-evals" in command for command in commands] == (
-        [True, False] if mode.startswith("benchmark-") else [False])
-    assert ("--all-evals" in commands[-1]) == mode.endswith("all")
-    assert "--evals-only" in commands[-1]
-    assert _scenario_values(commands[-1]) == ["fixed-seq-len", "agentic-coding"]
-    if mode.startswith("benchmark-"):
-        assert "--all-evals" not in commands[0]
+    expected_benchmarks = ([16] if trim else [16, 32, 64]) if mode.startswith("benchmark-") else []
+    assert [r["conc"] for r in output["single_node"].get("8k1k", [])] == expected_benchmarks
+    assert [r["conc"] for r in output["evals"]] == ([16, 32, 64] if mode.endswith("all") else [32, 64])
+    assert [r["conc"] for r in output["agentic_evals"]] == ([16, 32] if mode.endswith("all") else [32])
     assert output["changelog_metadata"]["base_ref"] == "base"
     for flag, value in entry_flags.items():
         assert output["changelog_metadata"]["entries"][0][flag] is value
 
 
-def test_overlapping_wildcard_scenarios_keep_priority_and_group_order(changelog_run):
-    entries = [
-        {"scenario-type": ["fixed-seq-len"]},
-        {"config-keys": ["config-*", "config-a"], "scenario-type": ["agentic-coding", "fixed-seq-len"]},
-        {"config-keys": ["config-b"], "all-evals": True, "scenario-type": ["agentic-coding"]},
-        {"config-keys": ["config-*"], "evals-only": True, "scenario-type": ["fixed-seq-len"]},
-    ]
-    _, commands = changelog_run(entries)
-    trace = [("benchmark" if "--no-evals" in c else "all" if "--all-evals" in c else "subset",
-              c[c.index("--config-keys") + 1:c.index("--config-files")], _scenario_values(c))
-             for c in commands]
-    assert trace == [
-        ("all", ["config-b"], ["agentic-coding"]),
-        ("benchmark", ["config-a"], ["fixed-seq-len"]),
-        ("subset", ["config-a"], ["fixed-seq-len"]),
-        ("benchmark", ["config-b"], []),
-        ("benchmark", ["config-a"], ["agentic-coding"]),
-        ("subset", ["config-b"], ["fixed-seq-len"]),
-        ("subset", ["config-a"], ["agentic-coding"]),
-    ]
+def test_overlapping_scenarios_preserve_eval_precedence_and_buckets(changelog_run):
+    output = changelog_run([
+        {"config-keys": ["*"], "scenario-type": ["fixed-seq-len"]},
+        {"config-keys": ["single"], "all-evals": True, "scenario-type": ["agentic-coding"]},
+        {"config-keys": ["single", "multi"], "scenario-type": ["agentic-coding", "fixed-seq-len"]},
+        {"config-keys": ["multi"], "all-evals": True, "scenario-type": ["fixed-seq-len"]},
+    ])
+    assert [r["conc"] for r in output["evals"]] == [32, 64]
+    assert [r["conc"] for r in output["agentic_evals"]] == [16, 32]
+    assert [r["conc"] for r in output["multinode_evals"]] == [[16, 32, 64]]
+    assert [r["conc"] for r in output["multinode_agentic_evals"]] == [[32]]
+    assert [r["conc"] for r in output["single_node"]["8k1k"]] == [16, 32, 64]
+    assert [r["conc"] for r in output["multi_node"]["8k1k"]] == [[16, 32, 64]]
+    assert [r["conc"] for r in output["single_node"]["agentic"]] == [16, 32]
+    assert [r["conc"] for r in output["multi_node"]["agentic"]] == [[16], [32]]
+    # Generated rows are independent across throughput and eval passes.
+    assert all(r.get("run-eval", False) is False for bucket in output["single_node"].values() for r in bucket)
+    assert all(r["eval-only"] is True for r in output["evals"] + output["agentic_evals"])
+
+
+@pytest.mark.parametrize("skip", [False, True])
+def test_no_evals_preserves_throughput_and_other_entries(changelog_run, skip):
+    output = changelog_run([{"no-evals": skip}, {"config-keys": ["multi"]}])
+    assert [r["conc"] for r in output["single_node"]["8k1k"]] == [16, 32, 64]
+    assert [r["conc"] for r in output["evals"]] == ([] if skip else [32, 64])
+    assert [r["conc"] for r in output["multinode_evals"]] == [[16, 32, 64]]
+    assert output["changelog_metadata"]["entries"][0]["no-evals"] is skip
+
+
+@pytest.mark.parametrize("flags", [{"evals-only": True}, {"all-evals": True}, {"eval-min-prefill-ep": 2}])
+def test_no_evals_rejects_conflicting_entry_options(changelog_run, flags):
+    with pytest.raises(ValueError, match="no-evals cannot be combined"):
+        changelog_run([{"no-evals": True, **flags}])
+
+
+@pytest.mark.parametrize("flag", ["--all-evals", "--evals-only"])
+def test_no_evals_rejects_conflicting_pr_modifiers(changelog_run, flag):
+    with pytest.raises(ValueError, match="no-evals entries cannot use"):
+        changelog_run([{"no-evals": True}], [flag])
 
 
 @pytest.mark.parametrize("cli_flags", [["--all-evals"], ["--evals-only"], ["--all-evals", "--evals-only"]])
 @pytest.mark.parametrize("trim", [False, True])
 def test_append_only_rejects_cli_eval_modifiers_before_generation(changelog_run, cli_flags, trim):
     with pytest.raises(ValueError, match="append-only sweeps cannot use"):
-        changelog_run([{"append-only": True}], cli_flags + (["--trim-conc"] if trim else []),
-                      lambda _: pytest.fail("invalid sweep must not invoke the generator"))
+        changelog_run([{"append-only": True}], cli_flags + (["--trim-conc"] if trim else []))
 
 
 @pytest.mark.parametrize("entries", [
-    [{"append-only": True}, {}],
-    [{"append-only": True, "all-evals": True}],
-    [{"append-only": True, "evals-only": True}],
-    [{"append-only": True, "eval-min-prefill-ep": 2}],
+    [{"append-only": True}, {}], [{"append-only": True, "all-evals": True}],
+    [{"append-only": True, "evals-only": True}], [{"append-only": True, "eval-min-prefill-ep": 2}],
 ])
 def test_append_only_rejects_mixed_or_entry_eval_modes(changelog_run, entries):
     with pytest.raises(ValueError, match="append-only"):
-        changelog_run(entries, generated=lambda _: pytest.fail("invalid sweep must not invoke the generator"))
-
-
-@pytest.mark.parametrize("stage", ["--no-evals", "--evals-only", "append-head", "append-base"])
-@pytest.mark.parametrize("failure", ["exit", "json"])
-def test_generator_failure_never_publishes_partial_matrix(changelog_run, monkeypatch, capsys, stage, failure):
-    error = subprocess.CalledProcessError(23, ["generator"], stderr="controlled generator failure")
-    append = stage.startswith("append-")
-    if append:
-        monkeypatch.setattr(process_changelog, "generation_inputs_at_ref", lambda _: nullcontext(
-            process_changelog.GenerationInputs(["base-config"], "base-generator", "base-runners")))
-
-    def generate(command):
-        revision_stage = "append-base" if command[1] == "base-generator" else "append-head"
-        if stage in command or (append and stage == revision_stage):
-            return error if failure == "exit" else "not-json"
-        return [_fixed_matrix_row(4), _fixed_matrix_row(8)]
-
-    with pytest.raises(subprocess.CalledProcessError if failure == "exit" else json.JSONDecodeError) as caught:
-        changelog_run([{"append-only": append}], generated=generate)
-    if failure == "exit":
-        assert caught.value is error
-    assert capsys.readouterr().out == ("controlled generator failure\n" if failure == "exit" else "")
+        changelog_run(entries)
 
 
 @pytest.mark.parametrize("keys,message", [
-    (["config-a", "missing"], "not found"),
-    (["config-a", "missing-*"], "No config keys matched"),
+    (["single", "missing"], "not found"), (["single", "missing-*"], "No config keys matched"),
 ])
-def test_invalid_key_after_valid_key_rejects_entire_selection(changelog_run, keys, message):
+def test_invalid_key_after_valid_key_rejects_entire_selection(changelog_run, keys, message, planning_repo):
+    # Key errors precede runner loading, even when that file is invalid.
+    (planning_repo[0] / "configs/runners.yaml").write_text("invalid\n")
     with pytest.raises(ValueError, match=message):
-        changelog_run([{"config-keys": keys}], generated=lambda _: pytest.fail("keys must resolve before generation"))
+        changelog_run([{"config-keys": keys}])
+
+
+@pytest.mark.parametrize("trim", [False, True])
+def test_append_only_main_runs_only_added_points_and_skips_evals(planning_repo, changelog_run, trim):
+    root, master, _ = planning_repo
+    git = lambda *args: subprocess.run(["git", *args], cwd=root, check=True, capture_output=True)
+    git("init", "-q")
+    git("config", "user.name", "Test")
+    git("config", "user.email", "test@example.com")
+    git("add", ".")
+    git("commit", "-qm", "base")
+    git("tag", "base")
+    master["single"]["scenarios"]["fixed-seq-len"][0]["search-space"][0]["conc-list"].append(128)
+    (root / "configs/nvidia-master.yaml").write_text(yaml.safe_dump(master, sort_keys=False))
+    output = changelog_run([{"append-only": True, "scenario-type": ["fixed-seq-len"]}],
+                           ["--trim-conc"] if trim else [])
+    rows = output["single_node"]["8k1k"]
+    assert [r["conc"] for r in rows] == [128]
+    assert len(rows[0]["recipe-fingerprint"]) == 64
+    assert output["evals"] == output["agentic_evals"] == []
+
+
+def test_generation_failure_does_not_publish_a_partial_matrix(planning_repo, changelog_run, capsys):
+    # Single-node generation succeeds before multinode scheduling fails.
+    (planning_repo[0] / "configs/runners.yaml").write_text("labels: {}\nhardware: {}\n")
+    with pytest.raises(subprocess.CalledProcessError):
+        changelog_run([{"config-keys": ["single", "multi"]}])
+    captured = capsys.readouterr()
+    assert "Cannot resolve gpus-per-node" in captured.out
+    assert '\"single_node\":' not in captured.out
 
 
 @pytest.mark.parametrize("threshold,expected", [
@@ -1170,3 +708,88 @@ def test_eval_prefill_ep_filter_preserves_single_node_and_order(threshold, expec
         {"label": "invalid", "prefill": {"ep": "bad"}},
     ]
     assert [row["label"] for row in process_changelog.filter_eval_rows_by_prefill_ep(rows, threshold)] == expected
+
+
+def test_current_plan_loads_inputs_once_and_uses_no_generator_process(planning_repo, monkeypatch):
+    import builtins
+    from collections import Counter
+    from infx.matrix.plan import build_plan
+
+    reads = Counter()
+    real_open = builtins.open
+
+    def counted_open(file, *args, **kwargs):
+        reads[str(file)] += 1
+        return real_open(file, *args, **kwargs)
+
+    monkeypatch.setattr(builtins, "open", counted_open)
+    monkeypatch.setattr(subprocess, "run", lambda *a, **kw: pytest.fail("current generation launched a child"))
+    entries = [{"config-keys": [key], "description": ["Controlled change"],
+                "pr-link": "https://github.com/SemiAnalysisAI/InferenceX/pull/1"}
+               for key in ("single", "multi")]
+    result = build_plan(entries, base_ref="base", head_ref="head").model_dump(by_alias=True, exclude_none=True)
+    assert [r["conc"] for r in result["single_node"]["8k1k"]] == [16, 32, 64]
+    assert result["multi_node"]["8k1k"][0]["node-count"] == 2
+    assert {path: reads[path] for path in (
+        "configs/amd-master.yaml", "configs/nvidia-master.yaml", "configs/runners.yaml",
+    )} == {"configs/amd-master.yaml": 1, "configs/nvidia-master.yaml": 1, "configs/runners.yaml": 1}
+
+
+def test_generation_api_preserves_inputs_and_returns_independent_nested_rows(planning_repo):
+    import copy
+    from infx.matrix.generate import generate_config_matrix
+
+    _, master, runners = planning_repo
+    master["multi"]["router"] = {"name": "fixture-router", "version": "1.0"}
+    original = copy.deepcopy((master, runners))
+    rows = generate_config_matrix(["multi"], master, runners, eval_mode="all")
+    assert [(r["conc"], r.get("eval-conc")) for r in rows] == [([16, 32, 64], None), ([16, 32], 32)]
+    rows[0]["router"]["name"] = "mutated"
+    rows[0]["prefill"]["tp"] = 999
+    rows[0]["conc"].append(999)
+    assert (master, runners) == original
+    repeated = generate_config_matrix(["multi"], master, runners, eval_mode="none")
+    assert repeated[0]["prefill"]["tp"] == 8
+    assert repeated[0]["conc"] == [16, 32, 64]
+
+
+def test_generation_api_rejects_unknown_eval_mode(planning_repo):
+    from infx.matrix.generate import generate_config_matrix
+    _, master, runners = planning_repo
+    with pytest.raises(ValueError, match="Unknown eval mode"):
+        generate_config_matrix(["single"], master, runners, eval_mode="typo")
+
+
+def test_plan_failure_after_throughput_generation_publishes_nothing(planning_repo, changelog_run, monkeypatch, capsys):
+    from infx.matrix import plan
+    generate = plan.generate_config_matrix
+
+    def fail_eval(*args, **kwargs):
+        if kwargs["eval_mode"] == "subset":
+            raise ValueError("controlled eval selection failure")
+        return generate(*args, **kwargs)
+
+    monkeypatch.setattr(plan, "generate_config_matrix", fail_eval)
+    with pytest.raises(subprocess.CalledProcessError):
+        changelog_run([{}])
+    output = capsys.readouterr().out
+    assert "controlled eval selection failure" in output
+    assert '"single_node":' not in output
+
+
+def test_plan_rejects_empty_changelog_before_reading_inputs():
+    from infx.matrix.plan import build_plan
+    with pytest.raises(ValueError, match="No valid YAML entries"):
+        build_plan([], base_ref="base", head_ref="head", config_files=["missing.yaml"])
+
+
+def test_generation_api_preserves_json_rejection_for_yaml_sets(planning_repo):
+    from infx.matrix.generate import generate_config_matrix
+    _, master, runners = planning_repo
+    # Pydantic accepts this YAML set as a list, but validation returns the raw
+    # config. The generator CLI has always rejected it at JSON serialization.
+    worker = master["multi"]["scenarios"]["fixed-seq-len"][0]["search-space"][0]["prefill"]
+    worker["additional-settings"] = {"A=1"}
+    validate_master_config(master)
+    with pytest.raises(TypeError, match="set is not JSON serializable"):
+        generate_config_matrix(["multi"], master, runners, eval_mode="none")
